+++ some-api-client/src/main/java/org/spring/boot/metrics/instances/some/api/client/service/DeleteService.java
@@ -1,0 +1,21 @@
+package org.spring.boot.metrics.instances.some.api.client.service;
+
+/**
+ * The interface Delete service.
+ *
+ * @param <S>  the type parameter
+ * @param <ID> the type parameter
+ * @author Alexander A. Kropotin
+ * @project some -api-client
+ * @created 2021 -07-21 20:19 <p>
+ */
+public interface DeleteService<S, ID> {
+
+    /**
+     * Delete s.
+     *
+     * @param id the id
+     * @return the s
+     */
+    S delete(ID id);
+}